+++ docs/main/changelog.rst
@@ -7,6 +7,11 @@ that were made in every particular version.
 From version 0.7.6 *Dependency Injector* framework strictly 
 follows `Semantic versioning`_
 
+4.26.0
+------
+- Add wiring by string id.
+- Improve error message for ``Dependency`` provider missing attribute.
+
 4.25.1
 ------
 - Amend docs and add another example for ``@containers.copy()`` decorator.